requirements.txt
python-telegram-bot[job-queue]==20.0
pandas==2.2.1
openpyxl==3.1.2
APScheduler~=3.9.1
pytz
